Add files via upload
Was having trouble bringing in a fork, so I did things in sort of a round about way. Unintentionally created a repository for 7B, so I went with it and created a new py file. Anyway, I had to change the interpreter as 2.7 doesn't support the "f" function, but arcpy does